examples/requirements-dev.txt
numpy==1.24.1
hydra-core==1.3.1
omegaconf==2.3.0
gym>=0.21.0,<1.0

git+https://github.com/Farama-Foundation/d4rl@71a9549f2091accff93eeff68f1f3ab2c0e0a288#egg=d4rl

wandb==0.13.9
torch